docs: rename demo4 and demo5
