+++ packages/hooks/src/useReactive/index.zh-CN.md
@@ -15,19 +15,19 @@ nav:
 
 ### 数组操作
 
-<code src="./demo/demo2.tsx"  />
+<code src="./demo/demo2.tsx" />
 
 ### 计算属性
 
-<code src="./demo/demo3.tsx"  />
-
-### 注意
-
-<code src="./demo/demo4.tsx" />
+<code src="./demo/demo3.tsx" />
 
 ### 元素属性
 
-<code src="./demo/demo5.tsx"  />
+<code src="./demo/demo4.tsx" />
+
+### 注意
+
+<code src="./demo/demo5.tsx" />
 
 ## API
